revamped search workflow
