+++ front-end/src/components/PreferencesForm/preferences-form.module.scss
@@ -4,7 +4,7 @@
     display: flex;
     flex-direction: column;
     align-items: center;
-    max-width: 40rem;
+    max-width: 50rem;
 }
 
 .selector {
