+++ front-end/src/pages/home/home.module.scss
@@ -21,7 +21,7 @@
     height: fit-content;
     display: grid;
     grid-template-columns: 2fr;
-    grid-template-rows: 2fr;
+    grid-template-rows: 3fr;
   }
 
   .headingContainer {
